feat: move createGameRoomCode function to utils folder
